requirements.txt
ddddocr>=1.0.6
pillow>=8.0.0
numpy>=1.19.0
fastapi>=0.95.0
uvicorn>=0.20.0
python-multipart>=0.0.5
requests>=2.28.0
aiohttp>=3.8.0
python-jose>=3.3.0
pydantic>=1.10.0
flask>=2.0.0
pyyaml>=6.0
